Clear browsing data on exit when instructed by the user.
Fixes brave/brave-browser#492

1. Adds a new "On exit" tab to chrome://settings/clearBrowserData dialog
with the same options as on the "Advanced" tab and a "Save" button. By
default all "On exit" options are turned off. When an option is modified
from the current state the Save button becomes enabled and allows the
user to save the "on exit" options. The options are saved to profile's
preferences: new *OnExit preferences have been added for each option.

2. Clear browsing data on exit functionality is triggered from Browser
destructor. The code checks if user selected any on exit options and if
the browser object being destroyed is the last one for the current
profile. If these conditions apply then a BrowsingDataRemover is called.

3. Adds BraveClearDataOnExitTest and BraveClearDataOnExitTwoBrowsersTest
browser tests that verify that:
- the removal of the browsing data is triggered only when an "on exit"
  preference is set;
- the removal flags are constrcuted correctly;
- the removal completes on exit;
- the removal is called only when the last active browser session for
  the profile is being closed (tests with OTR/Guest/multiple profiles
  being present).